manejo de errores Guardian Antidelito

+++ Python/src/GUImain/exceptionClasses/exceptionNoInt.py
@@ -11,7 +11,7 @@ class ExceptionNoInt(ErrorEnCampos):
 
     def checkValue(self, dato):
         try:
-            int(dato)
+            float(dato) 
         except:
             raise self
